[APDV-111] Moved FieldMapper
